requirements.txt
streamlit
django
djangorestframework
langchain
langchain-community
langchain-text-splitters
langchain-groq
faiss-cpu
sentence-transformers
python-dotenv
pytesseract
pdf2image
pillow
opencv-python
requests
